UI score input
- created parent class so api can support extract feedback and running the scorer in the same call
