+++ lib/BlisModels.js
@@ -14,4 +14,5 @@ tslib_1.__exportStar(require("./Session"), exports);
 tslib_1.__exportStar(require("./teach"), exports);
 tslib_1.__exportStar(require("./TrainDialog"), exports);
 tslib_1.__exportStar(require("./UserInput"), exports);
+tslib_1.__exportStar(require("./UI"), exports);
 //# sourceMappingURL=BlisModels.js.map
