+++ lib/Extract.d.ts
@@ -1,7 +1,6 @@
 import { LabeledEntity } from './TrainDialog';
 import { EntityMetaData } from './Entity';
 import { Metrics } from './Metrics';
-import { Memory } from './Memory';
 export declare class PredictedEntity extends LabeledEntity {
     score: number;
     metadata: EntityMetaData;
@@ -13,9 +12,4 @@ export declare class ExtractResponse {
     metrics: Metrics;
     packageId: string;
     constructor(init?: Partial<ExtractResponse>);
-}
-export declare class UIExtractResponse {
-    extractResponse: ExtractResponse;
-    memories: Memory[];
-    constructor(init?: Partial<UIExtractResponse>);
 }
