+++ lib/Extract.js
@@ -5,7 +5,6 @@ var json_typescript_mapper_1 = require("json-typescript-mapper");
 var TrainDialog_1 = require("./TrainDialog");
 var Entity_1 = require("./Entity");
 var Metrics_1 = require("./Metrics");
-var Memory_1 = require("./Memory");
 var PredictedEntity = (function (_super) {
     tslib_1.__extends(PredictedEntity, _super);
     function PredictedEntity(init) {
@@ -53,21 +52,4 @@ var ExtractResponse = (function () {
     return ExtractResponse;
 }());
 exports.ExtractResponse = ExtractResponse;
-var UIExtractResponse = (function () {
-    function UIExtractResponse(init) {
-        this.extractResponse = undefined;
-        this.memories = undefined;
-        Object.assign(this, init);
-    }
-    tslib_1.__decorate([
-        json_typescript_mapper_1.JsonProperty({ clazz: ExtractResponse, name: 'extractResponse' }),
-        tslib_1.__metadata("design:type", ExtractResponse)
-    ], UIExtractResponse.prototype, "extractResponse", void 0);
-    tslib_1.__decorate([
-        json_typescript_mapper_1.JsonProperty({ clazz: Memory_1.Memory, name: 'memories' }),
-        tslib_1.__metadata("design:type", Array)
-    ], UIExtractResponse.prototype, "memories", void 0);
-    return UIExtractResponse;
-}());
-exports.UIExtractResponse = UIExtractResponse;
 //# sourceMappingURL=Extract.js.map
