+++ lib/Score.js
@@ -4,7 +4,6 @@ var tslib_1 = require("tslib");
 var json_typescript_mapper_1 = require("json-typescript-mapper");
 var Metrics_1 = require("./Metrics");
 var Action_1 = require("./Action");
-var Memory_1 = require("./Memory");
 var ScoreInput = (function () {
     function ScoreInput(init) {
         this.filledEntities = undefined;
@@ -111,26 +110,4 @@ var ScoreResponse = (function () {
     return ScoreResponse;
 }());
 exports.ScoreResponse = ScoreResponse;
-var UIScoreResponse = (function () {
-    function UIScoreResponse(init) {
-        this.scoreResponse = undefined;
-        this.scoreInput = undefined;
-        this.memories = undefined;
-        Object.assign(this, init);
-    }
-    tslib_1.__decorate([
-        json_typescript_mapper_1.JsonProperty({ clazz: ScoreResponse, name: 'scoreResponse' }),
-        tslib_1.__metadata("design:type", ScoreResponse)
-    ], UIScoreResponse.prototype, "scoreResponse", void 0);
-    tslib_1.__decorate([
-        json_typescript_mapper_1.JsonProperty({ clazz: ScoreInput, name: 'scoreInput' }),
-        tslib_1.__metadata("design:type", ScoreInput)
-    ], UIScoreResponse.prototype, "scoreInput", void 0);
-    tslib_1.__decorate([
-        json_typescript_mapper_1.JsonProperty({ clazz: Memory_1.Memory, name: 'memories' }),
-        tslib_1.__metadata("design:type", Array)
-    ], UIScoreResponse.prototype, "memories", void 0);
-    return UIScoreResponse;
-}());
-exports.UIScoreResponse = UIScoreResponse;
 //# sourceMappingURL=Score.js.map
