+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "blis-models",
-  "version": "0.29.0",
+  "version": "0.30.2",
   "description": "Models for Blis",
   "main": "lib/BlisModels.js",
   "typings": "lib/BlisModels",
