+++ src/Score.ts
@@ -97,23 +97,3 @@ export class ScoreResponse
         (<any>Object).assign(this, init);
     }
 }
-
-export class UIScoreResponse
-{
-    @JsonProperty({clazz: ScoreResponse, name: 'scoreResponse'})
-    public scoreResponse : ScoreResponse;
-
-    @JsonProperty({clazz: ScoreInput, name: 'scoreInput'})
-    public scoreInput : ScoreInput;
-
-    @JsonProperty({clazz: Memory, name: 'memories'})
-    public memories : Memory[];
-
-    public constructor(init?:Partial<UIScoreResponse>)
-    {
-        this.scoreResponse = undefined;
-        this.scoreInput = undefined;
-        this.memories = undefined;
-        (<any>Object).assign(this, init);
-    }
-}
